requirements.txt
argcomplete>=1.8.2
boto3>=1.4.4
boto
botocore>=1.5.13
-e git+https://github.com/capitalone/cloud-custodian.git#egg=c7n
docutils>=0.12
EasyProcess
functools32==3.2.3-2
futures==3.0.5
ipaddress>=1.0.16
Jinja2>=2.8
jmespath>=0.9.0
jsonschema>=2.5.1
ldap3>=2.1.1
Markup
MarkupSafe==0.23
python-dateutil>=2.5.2
python-memcached==1.58
PyYAML>=3.12
six==1.10.0
vcversioner==2.16.0.0
